# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VORA API",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/drivers": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List drivers by KYC status",
        "description": "Review queue. Returns no phone numbers and no decrypted document references: approving KYC needs to know which documents exist, not what they say, so an admin session compromise leaks neither.",
        "operationId": "list_drivers_api_v1_admin_drivers_get",
        "parameters": [
          {
            "name": "kyc_status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/KycStatus",
              "default": "pending"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriverAdminView"
                  },
                  "title": "Response List Drivers Api V1 Admin Drivers Get"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or schema failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but role-forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Semantically invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Carries Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented yet (Phase 0 stub)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/drivers/{driver_id}/kyc": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Approve, reject or suspend a driver",
        "description": "Moves the driver through the KYC state machine. Verification is refused while any required document is missing, and that check lives in the service rather than in a console, so no second interface can bypass it. Every decision is written to the audit log with the reviewer's id: Law 2024/017 requires an annual security report, and 'who approved this driver' is exactly what it has to answer.",
        "operationId": "decide_kyc_api_v1_admin_drivers__driver_id__kyc_post",
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Driver Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KycDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverAdminView"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or schema failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but role-forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Semantically invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Carries Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented yet (Phase 0 stub)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DriverAdminView": {
        "properties": {
          "driver_id": {
            "type": "string",
            "format": "uuid",
            "title": "Driver Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "kyc_status": {
            "$ref": "#/components/schemas/KycStatus"
          },
          "submitted": {
            "items": {
              "$ref": "#/components/schemas/KycDocumentKind"
            },
            "type": "array",
            "title": "Submitted"
          },
          "missing": {
            "items": {
              "$ref": "#/components/schemas/KycDocumentKind"
            },
            "type": "array",
            "title": "Missing"
          },
          "is_online": {
            "type": "boolean",
            "title": "Is Online"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "driver_id",
          "user_id",
          "display_name",
          "kyc_status",
          "submitted",
          "missing",
          "is_online"
        ],
        "title": "DriverAdminView",
        "description": "What an admin sees about a driver under review.\n\nNo phone number and no decrypted document reference. An admin approving\nKYC needs to know which documents exist and what state the account is in;\nthey do not need the identity numbers themselves, and not returning them\nmeans an admin session compromise does not leak them either."
      },
...
      "ErrorDetailModel": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetailModel"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailModel"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The §6 envelope. Every non-2xx response in this API has this shape."
      },
...
      "KycDecisionRequest": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/KycStatus"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "Required when rejecting, so the driver can correct and resubmit."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "KycDecisionRequest",
        "description": "Admin decision on a driver's KYC."
      },
      "KycDocumentKind": {
        "type": "string",
        "enum": [
          "cni",
          "driving_licence",
          "vehicle_registration",
          "vehicle_photo",
          "driver_photo"
        ],
        "title": "KycDocumentKind"
      },
...
      "KycStatus": {
        "type": "string",
        "enum": [
          "pending",
          "verified",
          "rejected",
          "suspended"
        ],
        "title": "KycStatus"
      },
...
    }
  }
}